feature/ui layer (#3)
* Do NOT load feed on init

Before the view is loaded, otherwise it would not
be visible.

* Lint files

* Load feed on viewDidLoad

* Replace spy with 'FeedLoader'

* Refactor SUT to factory method

* Reload on pull to refresh

* Show loading indicator on viewDidLoad

* Hide loading indicator on load completion

* Pull to refresh toggles loading indicator

* Decouple implementation details from tests

Pull-to-refresh and UIRefreshControl should not
break the tests if we decide to initate a refresh
via a button instead, or use a custom refresh control.

* Combine relevant tests to eliminate temporal coupling bugs

* Move 'FeedViewController' to production

* Render loaded feed on successful load

* Preserve current feed state on error

When the feed has successfully loaded, and then it
fails when reloading, makes that the initial loaded
feed is still visibile.

* Hide loading indicator on completion

* Load image url when image becomes visible

* Cancel image loading when image is hidden

* Cancel image request when no longer visible

Using a separate protocol 'FeedImageDataTask' as a return
type, to respect the Interface Segregation Principle'.

This moves the responsibility of state management to the
clients of the 'FeedImageDataLoader' protocol

* Feed Image loading indicator is visible when loading

* Separate test file with Mark tags

* Render loaded images from URL

* 'FeedImageView.retryButton' is visible on error

* Show retry button is image data is invalid

* Retry failed image load on retry action

* Preload image URL when image is near visible

* Cancel reloading when image is no longer visible

* Organize folder hierarchy for MVC

* Extract "Refresh/Load" to 'FeedRefreshController'

Includes the creation and configuration on the 'UIRefreshControl'
and the refresh logic contained in the 'FeedLoader'

* Extract cell logic to 'FeedImageCellController'

Removing cell creation, configuration, and
loading/preloading  logic

* Extract creation to 'FeedUIComposer'

Dependency creation and composition logic responsibility
is moved away from the 'FeedViewController'

* Extract adapter pattern into a separate function to clarify intent

* Delete empty docs

* Move loading state management to ViewModel

Makes the RefreshController platform agnostic, by decoupling
the model dependencies. Now the 'FeedRefreshController' acts as
a binding/binder between the 'View' and the 'ViewModel'

* Remove mutable state from 'FeedViewModel'

* Set development team (new machine)

* Extract 'EssentialFeed' model logic

* Remember to commit - Because I can't recall 4 days ago

* Extract UIKit dependency from ViewModel

* Remove 'EssentialFeed' dependency from 'FeedViewController'

* Rename 'FeedImageViewModel' removing 'Cell'

* Add 'FeedPresenter' to replace 'FeedViewModel'

Presenter holds presentation logic, consisting of
the loading state and loaded feed.

* Move memory management to Composition layer

The 'FeedPresenter' shouldn't know about or handle
the dependencies life cycle

* Delete 'FeedViewModel'

No longer used. Replaced by the 'FeedPresenter' to
demonstrate the MVP architecture

* Add Presentable ViewModels as pure data

Clarifys the communication between Presentation and UI

* Decouple 'FeedRefreshViewController' from 'FeedPresenter'

* Decouple 'FeedPresenter' from 'FeedLoader'

via adapter in the composition layer

* Replace closure event handler with delegate

For demonstration purposes of how easy it is to
refactor a closure to a protocol

* Swap property injection to constructor injection

Guarantee's the 'FeedPresenter' has its required properties,
to prevent leaking composition details.

Co-authored-by: Jason Kuan <jsonkuan@jasons-mac-mini.lan>